Sketch out the rough shape of our app
As we learn more tailwind we can start to sketch out the shape of our
layout a bit more...

This also adds some hardcoded items. In the future we will dynamically
add / persist / update the inventory items based on user interactions
but for right now we are just focusing on using our new tailwind library
to build out the shape of our goodies! The next stage will be adding
more and more functionality to this!