Get Delete working and add an add-item helper
We're basically making a todo list with a fake db so lets do that.
We wrap adding a new item in a helper to generate an id so we can look
up our item in a sorted list and we add a delete function to remove that
item from the list by id. This is better than using a vector or list
because we don't need to iterate through it each time and can instead
use the index to determine what should change.

We also define a counter to increment our ids each time we add a new one
so we can get rid of the right ones! Without worrying about conflicting
ids!